Move compiler classes into compiler module
