+++ modules/compiler/src/main/java/script/control/PhaseAware.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nextflow.lsp.compiler;
+package nextflow.script.control;
 
 /**
  *
