+++ modules/compiler/src/main/java/script/control/Phases.java
@@ -13,15 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nextflow.lsp.compiler;
-
-import org.codehaus.groovy.ast.ASTNode;
+package nextflow.script.control;
 
 /**
  *
  * @author Ben Sherman <bentshermann@gmail.com>
  */
-public interface RelatedInformationAware {
-    String getOtherMessage();
-    ASTNode getOtherNode();
+public class Phases {
+    public static final int SYNTAX = 1;
+    public static final int NAME_RESOLUTION = 2;
+    public static final int INCLUDE_RESOLUTION = 3;
+    public static final int TYPE_INFERENCE = 4;
+
+    public static final int ALL = TYPE_INFERENCE;
 }
