+++ modules/language-server/src/main/java/nextflow/lsp/services/config/ConfigAstCache.java
@@ -32,10 +32,10 @@ import nextflow.lsp.ast.ASTNodeCache;
 import nextflow.lsp.ast.ASTParentVisitor;
 import nextflow.lsp.compiler.Compiler;
 import nextflow.lsp.compiler.LanguageServerErrorCollector;
-import nextflow.lsp.compiler.PhaseAware;
-import nextflow.lsp.compiler.Phases;
 import nextflow.lsp.file.FileCache;
-import org.codehaus.groovy.ast.ASTNode;;
+import nextflow.script.control.PhaseAware;
+import nextflow.script.control.Phases;
+import org.codehaus.groovy.ast.ASTNode;
 import org.codehaus.groovy.ast.Parameter;
 import org.codehaus.groovy.ast.expr.BinaryExpression;
 import org.codehaus.groovy.ast.expr.BitwiseNegationExpression;
