+++ modules/language-server/src/main/java/nextflow/lsp/services/config/ResolveIncludeVisitor.java
@@ -24,8 +24,8 @@ import java.util.Set;
 import nextflow.config.ast.ConfigIncludeNode;
 import nextflow.config.ast.ConfigNode;
 import nextflow.config.ast.ConfigVisitorSupport;
-import nextflow.lsp.compiler.PhaseAware;
-import nextflow.lsp.compiler.Phases;
+import nextflow.script.control.PhaseAware;
+import nextflow.script.control.Phases;
 import org.codehaus.groovy.ast.ASTNode;
 import org.codehaus.groovy.ast.expr.ConstantExpression;
 import org.codehaus.groovy.control.SourceUnit;
